updated if/else to include error for past dates.
