+++ src/client/index.js
@@ -1,5 +1,6 @@
 import { processInfo } from "./js/app.js";
 import { calcDate } from "./js/calculateDate.js";
+import { getCurrentWeather, getFutureWeather } from "./js/weatherApiCalls";
 
 // console.log(testFuntion)
 // alert('HELLO WORLD')
@@ -11,5 +12,4 @@ import'./styles/fonts.scss'
 
 // document.getElementById("submit").addEventListener("click", processInfo);
 
-export { processInfo}
-export {calcDate}
+export { processInfo, calcDate, getFutureWeather, getCurrentWeather}
